+++ components/frontend/src/widgets/Button.js
@@ -1,15 +1,13 @@
 import React, { useRef, useState } from 'react';
-import { PropTypes } from 'prop-types';
+import PropTypes from 'prop-types';
 import { Icon, Input } from 'semantic-ui-react';
 import { Button, Dropdown, Label, Popup } from '../semantic_ui_react_wrappers';
-import { get_report_pdf } from '../api/report';
-import { registeredURLSearchParams } from '../hooks/url_search_query';
 import { showMessage } from '../widgets/toast';
 import { ItemBreadcrumb } from './ItemBreadcrumb';
 
 export function ActionButton(props) {
     const { action, disabled, icon, item_type, floated, fluid, popup, position, ...other } = props;
-    const label = `${action} ${item_type}`;
+    const label = `${action} ${item_type}`
     // Put the button in a span so that a disabled button can still have a popup
     // See https://github.com/Semantic-Org/Semantic-UI-React/issues/2804
     const button = (
@@ -144,48 +142,6 @@ export function DeleteButton(props) {
             popup={`Delete this ${props.item_type}. Careful, this can only be undone by a system administrator!`}
             position='top right'
             {...props}
-        />
-    )
-}
-
-function download_pdf(report_uuid, query_string, callback) {
-    const reportId = report_uuid ? `report-${report_uuid}` : "reports-overview"
-    get_report_pdf(report_uuid, query_string)
-        .then(response => {
-            if (response.ok === false) {
-                showMessage("error", "PDF rendering failed", "HTTP code " + response.status + ": " + response.statusText)
-            } else {
-                let url = window.URL.createObjectURL(response);
-                let a = document.createElement('a');
-                a.href = url;
-                const now = new Date();
-                const local_now = new Date(now.getTime() - (now.getTimezoneOffset() * 60000));
-                a.download = `Quality-time-${reportId}-${local_now.toISOString().split(".")[0]}.pdf`;
-                a.click();
-            }
-        }).finally(() => callback());
-}
-
-export function DownloadAsPDFButton({ report_uuid }) {
-    const [loading, setLoading] = useState(false);
-    // Make sure the report_url contains only registered query parameters
-    const query = registeredURLSearchParams();
-    const queryString = query.toString() ? ("?" + query.toString()) : ""
-    query.set("report_url", window.location.origin + window.location.pathname + queryString + window.location.hash);
-    const itemType = report_uuid ? "report" : "reports overview"
-    return (
-        <ActionButton
-            action='Download'
-            icon="file pdf"
-            item_type={`${itemType} as PDF`}
-            loading={loading}
-            onClick={() => {
-                if (!loading) {
-                    setLoading(true);
-                    download_pdf(report_uuid, `?${query.toString()}`, () => { setLoading(false) })
-                }
-            }}
-            popup={`Generate a PDF version of the ${itemType} as currently displayed. This may take some time.`}
         />
     )
 }
